Clean dockers and database between each tests.
WARNING: regarding docker we've got some a classpath clash regarding docker client and testcontainers. Testcontainers rely to a shaded version of docker client incompatible with the one used. It seems that the order of declaration in pom.xml is very important !!!

WARNING: Injecting Agroal did not work. I add to define it manually...